requirements.txt
pandas~=1.4.3
nltk~=3.7
unbabel-comet~=2.0.0
modernmt~=1.1.0
openai~=0.27.2
tensorRT~=8.6.0
scikit-learn~=1.1.3
torch~=1.13.1
requests~=2.28.2
transformers~=4.27.2
tqdm~=4.64.0
peft~=0.2.0
sentence-transformers==2.2.2
hydra-core==1.3.2
omegaconf==2.3.0
tiktoken==0.3.3
lightning==2.0.1
wandb==0.14.0
deepspeed==0.8.3